Rename reference-line-utils to reference-line, fix perpendicular offset for left/right sides
- Renamed reference-line-utils.ts to reference-line.ts
- Fixed left/right annotation offset to use perpendicular direction for angled lines
  (previously only applied horizontal offset, ignoring line angle)
- Simplified tests using test.each and toMatchObject patterns
- Removed non-null assertions in favor of proper if guards
- Updated all imports to use new module path
