+++ src/lib/plot/index.ts
@@ -12,8 +12,8 @@ export { default as BarPlotControls } from './BarPlotControls.svelte'
 export { default as ColorBar } from './ColorBar.svelte'
 export { default as ColorScaleSelect } from './ColorScaleSelect.svelte'
 export { default as ElementScatter } from './ElementScatter.svelte'
-export { default as FillArea } from './FillArea.svelte'
 export * from './fill-utils'
+export { default as FillArea } from './FillArea.svelte'
 export { default as Histogram } from './Histogram.svelte'
 export { default as HistogramControls } from './HistogramControls.svelte'
 export * from './layout'
@@ -21,10 +21,10 @@ export { default as Line } from './Line.svelte'
 export { default as PlotControls } from './PlotControls.svelte'
 export { default as PlotLegend } from './PlotLegend.svelte'
 export { default as PlotTooltip } from './PlotTooltip.svelte'
+export * from './reference-line'
 export { default as ReferenceLine } from './ReferenceLine.svelte'
 export { default as ReferenceLine3D } from './ReferenceLine3D.svelte'
 export { default as ReferencePlane } from './ReferencePlane.svelte'
-export * from './reference-line-utils'
 export * from './scales'
 export { default as ScatterPlot } from './ScatterPlot.svelte'
 export { default as ScatterPlot3D } from './ScatterPlot3D.svelte'
